Fix channel filter
The channel filtering wasn't working, because the channel name wasn't
passed from Murmur as part of the message events.

Fix this by retrieving the list of channels at startup, finding the ones
that are referenced in the config and constructing a list of channel
ids.

This is used to filter mumble messages so that only those to the
specified channels are passed through to Matrix.

We also pass the same list of channels back to Murmur when relaying
Matrix messages, so that they only go to the specified channels.